Backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
openai>=1.0.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0
gTTS>=2.5.0
reportlab>=4.0.0
Pillow>=10.0.0
